# docs/swagger.yaml
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  schema.AcceptAnswerReq:
    properties:
      answer_id:
        type: string
      question_id:
        maxLength: 30
        type: string
    required:
    - question_id
    type: object
...
  schema.AdminUpdateAnswerStatusReq:
    properties:
      answer_id:
        type: string
      status:
        enum:
        - available
        - deleted
        type: string
    required:
    - answer_id
    - status
    type: object
...
  schema.AnswerAddReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      content:
        maxLength: 65535
        minLength: 6
        type: string
      question_id:
        type: string
    required:
    - content
    type: object
  schema.AnswerUpdateReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      content:
        maxLength: 65535
        minLength: 6
        type: string
      edit_summary:
        type: string
      id:
        type: string
      question_id:
        type: string
      title:
        type: string
    required:
    - content
    type: object
...
  schema.RecoverAnswerReq:
    properties:
      answer_id:
        type: string
    required:
    - answer_id
    type: object
...
  schema.RemoveAnswerReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      id:
        type: string
    required:
    - id
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/admin/api/answer/status:
    put:
      consumes:
      - application/json
      description: update answer status
      parameters:
      - description: AdminUpdateAnswerStatusReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AdminUpdateAnswerStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update answer status
      tags:
      - admin
...
  /answer/api/v1/answer:
    delete:
      consumes:
      - application/json
      description: delete answer
      parameters:
      - description: answer
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RemoveAnswerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: delete answer
      tags:
      - api-answer
    post:
      consumes:
      - application/json
      description: Insert Answer
      parameters:
      - description: AnswerAddReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AnswerAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Insert Answer
      tags:
      - api-answer
    put:
      consumes:
      - application/json
      description: Update Answer
      parameters:
      - description: AnswerUpdateReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AnswerUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update Answer
      tags:
      - api-answer
  /answer/api/v1/answer/acceptance:
    post:
      consumes:
      - application/json
      description: Accepted
      parameters:
      - description: AcceptAnswerReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AcceptAnswerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Accepted
      tags:
      - api-answer
  /answer/api/v1/answer/info:
    get:
      consumes:
      - application/json
      description: Get Answer
      parameters:
      - default: "1"
        description: Answer TagID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get Answer
      tags:
      - api-answer
  /answer/api/v1/answer/page:
    get:
      consumes:
      - application/json
      description: AnswerList <br> <b>order</b> (default or updated)
      parameters:
      - description: question_id
        in: query
        name: question_id
        required: true
        type: string
      - description: order
        in: query
        name: order
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: page_size
        in: query
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: AnswerList
      tags:
      - api-answer
  /answer/api/v1/answer/recover:
    post:
      consumes:
      - application/json
      description: recover deleted answer
      parameters:
      - description: answer
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RecoverAnswerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: recover answer
      tags:
      - Answer
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"